Fixed Seeders for insert in local or production data

+++ database/seeds/AccessControlApiTableSeeder.php
@@ -7,7 +7,10 @@ class AccessControlApiTableSeeder extends Seeder
 {
     public function run()
     {
-	    DB::table('access_control_apis')->truncate();
-        factory(AccessControlApi::class, 50)->create();
+	    if (getenv('APP_ENV') === 'local')
+	    {
+		    DB::table('access_control_apis')->truncate();
+		    factory(AccessControlApi::class, 50)->create();
+	    }
     }
 }
